requirements.txt
Flask==3.0.0
line-bot-sdk==2.5.0
requests==2.31.0
langdetect==1.0.9
